requirements.txt
feedparser==6.0.10
psycopg2-binary==2.9.9
